requirements.txt
python-dotenv==1.0.1
pydantic==2.2.1
gradio==5.39.0
openai==1.68.2
agents==0.1.0
aiohttp==3.8.5
typing-extensions==4.7.1
